scripts/requirements.txt
sentence-transformers==2.2.2
supabase==2.3.0
python-dotenv==1.0.0
tqdm==4.66.1
torch>=2.0.0